Add files via upload
This is the main code that can be used to exploit tools of agent.ai to create a website (created by the AI Agent we had built) and save it to our PC then host it on github. There can be potential clashes in this code that's why I have added one more code that can do half part properly.